Move, and rename, the `src/scripting_api/quickjs-sandbox.js` file to `src/pdf.sandbox.js`
The current location feels somewhat strange, and also inconsistent with the existing way that bundling is done.

Finally, add the version/build numbers at the top of the *built* `pdf.sandbox.js` files, since all other built files include that information given that it's often helpful to be able to easily determine the *exact* version.